Fixed some lsp issues + added type hints.

+++ icy_ppe/src/vm/expressions/predefined_functions.rs
@@ -564,9 +564,7 @@ pub fn valtime(vm: &mut VirtualMachine, params: &[VariableValue]) -> Res<Variabl
     panic!("TODO")
 }
 pub fn pcbnode(vm: &mut VirtualMachine, params: &[VariableValue]) -> Res<VariableValue> {
-    Ok(VariableValue::new_int(
-        vm.icy_board_data.pcb_data.node_num as i32,
-    ))
+    Ok(VariableValue::new_int(vm.icy_board_data.pcb_data.node_num))
 }
 
 pub fn readline(vm: &mut VirtualMachine, params: &[VariableValue]) -> Res<VariableValue> {
